Simplified containers and components by using redux-connect
Here we are no longer passing the store down from component to
component, which can go on and on as we start having more components
and deeper nesting.

Instead we have a simple container that is only in charge of arranging
the data the component directly within it needs (by accessing the
store), and passing it as props and nothing more to the component.

In fact that container is now so simple (once I also moved the JSX from
it to the presentational componenet) that it is no longer even
dependent on React. All it knows how to do is wrap the component within
it and pass it the data it requires.

This then allows the component inside to be uncaring of how the store
looks, and how to retrieve data from it. It just knows it has the data
it needs.

In fact... well, I'm going to save this refactor for the next commit.