Refactored how BadgeContainer passes data to Badge
Previously Badge had to be aware of how the state stores data as an
array of ids and an object containing badges by id. This complexity
really isn't any of Badge's concern - it can be resolved inside the
container, which can then simply pass the data Badge needs in an easily
digestible format an array of badges with their details.

This then allows us to simplify Badge significantly.